% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_landcover.R
\name{calculate_ndvi}
\alias{calculate_ndvi}
\title{Calculer le NDVI}
\usage{
calculate_ndvi(red = NULL, nir = NULL, n_dates = 3, scale_factor = 1)
}
\arguments{
\item{red}{Raster de la bande rouge (ou NULL pour données simulées)}

\item{nir}{Raster de la bande proche infrarouge (ou NULL pour données simulées)}

\item{n_dates}{Nombre de dates si simulation (défaut: 3)}

\item{scale_factor}{Facteur d'échelle à appliquer (défaut: 1.0)}
}
\value{
Un SpatRaster de NDVI (valeurs entre -1 et 1)
}
\description{
Calcule l'indice de végétation par différence normalisée (NDVI) à partir
des bandes rouge et proche infrarouge d'une image satellitaire.

NDVI = (NIR - Rouge) / (NIR + Rouge)
}
\examples{
# NDVI simulé
ndvi <- calculate_ndvi(n_dates = 4)
terra::plot(ndvi, main = "NDVI", col = grDevices::hcl.colors(100, "Greens"))

# Avec des rasters réels:
# ndvi <- calculate_ndvi(red = rast("bande_rouge.tif"), nir = rast("bande_nir.tif"))

}
