% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_functions.R
\name{clean_lst_data}
\alias{clean_lst_data}
\title{Nettoyer les données LST}
\usage{
clean_lst_data(
  lst_raster,
  min_temp = -10,
  max_temp = 65,
  cloud_threshold = NULL,
  interpolate_na = TRUE,
  smooth = NULL
)
}
\arguments{
\item{lst_raster}{Un objet SpatRaster contenant les données LST}

\item{min_temp}{Température minimale valide en °C (défaut: -10)}

\item{max_temp}{Température maximale valide en °C (défaut: 65)}

\item{cloud_threshold}{Seuil de détection nuages (défaut: NULL, désactivé)}

\item{interpolate_na}{Interpoler les valeurs manquantes (défaut: TRUE)}

\item{smooth}{Lisser le raster (fenêtre de lissage en cellules, défaut: NULL)}
}
\value{
Un SpatRaster nettoyé
}
\description{
Nettoie un raster LST en supprimant les pixels nuageux, corrigeant les
valeurs aberrantes et interpolant les valeurs manquantes.
}
\examples{
lst <- simulate_raster_data(type = "lst", n_dates = 3)
lst_clean <- clean_lst_data(lst, min_temp = 5, max_temp = 55)
terra::plot(lst_clean[[1]], main = "LST nettoyee")

}
