% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_report.R
\name{compare_lst_vs_ndvi}
\alias{compare_lst_vs_ndvi}
\title{Comparer LST et NDVI}
\usage{
compare_lst_vs_ndvi(
  lst_raster,
  ndvi_raster,
  layer_index = 1,
  n_sample = 1000,
  plot = TRUE
)
}
\arguments{
\item{lst_raster}{SpatRaster de LST}

\item{ndvi_raster}{SpatRaster de NDVI}

\item{layer_index}{Index de couche à utiliser (défaut: 1)}

\item{n_sample}{Nombre de pixels à échantillonner (défaut: 1000)}

\item{plot}{Afficher le graphique (défaut: TRUE)}
}
\value{
Une liste avec: correlation, p_value, plot (ggplot), data
}
\description{
Analyse la relation entre la température de surface (LST) et le NDVI.
Produit un scatterplot avec régression et calcule le coefficient de corrélation.
}
\examples{
lst  <- simulate_raster_data(type = "lst",  n_dates = 3)
ndvi <- simulate_raster_data(type = "ndvi", n_dates = 3)
result <- compare_lst_vs_ndvi(lst, ndvi, layer_index = 1)
cat("Correlation LST-NDVI:", result$correlation, "\n")

}
