% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_landcover.R
\name{extract_microclimate_features}
\alias{extract_microclimate_features}
\title{Extraire les variables microclimatiques}
\usage{
extract_microclimate_features(
  lst_raster,
  ndvi_raster = NULL,
  landcover_raster = NULL,
  points = NULL,
  n_sample = 500,
  layer_index = 1
)
}
\arguments{
\item{lst_raster}{SpatRaster de LST}

\item{ndvi_raster}{SpatRaster de NDVI (optionnel)}

\item{landcover_raster}{SpatRaster d'occupation du sol (optionnel)}

\item{points}{Objet sf de points d'extraction (optionnel, sinon grille régulière)}

\item{n_sample}{Nombre de points d'échantillonnage si pas de points fournis (défaut: 500)}

\item{layer_index}{Index de couche LST à utiliser (défaut: 1)}
}
\value{
Un data.frame avec variables microclimatiques
}
\description{
Extrait et agrège les variables microclimatiques (LST, NDVI, altitude,
occupation du sol) pour créer un dataframe d'analyse.
}
\examples{
lst  <- simulate_raster_data(type = "lst",  n_dates = 1)
ndvi <- simulate_raster_data(type = "ndvi", n_dates = 1)
lc   <- import_landcover()
features <- extract_microclimate_features(lst, ndvi, lc)
head(features)

}
